a lot of parameters can be configured in the env file
