+++ events/messageCreate.js
@@ -1,6 +1,7 @@
 const { Events } = require('discord.js');
 const { createEmbedMessage } = require('../utils');
 const ollamaService = require('../services/ollamaService');
+require('dotenv').config();
 
 module.exports = {
     name: Events.MessageCreate,
@@ -11,17 +12,15 @@ module.exports = {
 
         if (!promptMessage) return;
 
-        const fetchedMessages = await message.channel.messages.fetch({ limit: 11 });
+        const fetchMessageAmount = Math.abs(parseInt(process.env.OLLAMA_MODEL_FETCH_MESSAGE_AMOUNT, 10)) || 10;
+        const fetchedMessages = await message.channel.messages.fetch({ limit: fetchMessageAmount });
         const messageHistory = Array.from(fetchedMessages.values());
 
         const lastMessage = messageHistory[1];
 
         if (lastMessage && lastMessage.author.id === message.author.id) return;
 
-        let messages = createRoleMessages(messageHistory);
-        messages = createSystemMessage(messages, 'You are a helpful AI Assistant');
-
-        const response = await ollamaService.chat(messages);
+        const response = await ollamaService.chat(messageHistory);
 
         if (response) {
             await message.reply({ embeds: [createEmbedMessage(message.guild, response)] });
@@ -30,19 +29,3 @@ module.exports = {
         }
     },
 };
-
-function createRoleMessages(messages) {
-    return messages.map(m => {
-        if (m.author.bot) {
-            return { 'role': 'assistant', 'content': `${m.embeds[0].description}` };
-        } else {
-            return { 'role': 'user', 'content': `${m.content}` };
-        }
-
-    }).reverse();
-}
-
-function createSystemMessage(messages, message) {
-    messages.unshift({ 'role': 'system', 'content': message });
-    return messages;
-}
