+++ services/ollamaService.js
@@ -5,6 +5,10 @@ class OllamaService {
     constructor() {
         this.networkManager = new NetworkManager(`http://ollama:${process.env.OLLAMA_PORT}`);
         this.model = process.env.OLLAMA_MODEL;
+        this.maxInputTokenLength = parseInt(process.env.OLLAMA_MODEL_MAX_INPUT_TOKEN_LENGTH, 10);
+        this.maxOutputTokenLength = parseInt(process.env.OLLAMA_MODEL_MAX_OUTPUT_TOKEN_LENGTH, 10);
+        this.defaultPrompt = process.env.OLLAMA_MODEL_DEFAULT_PROMPT;
+        this.temperature = parseFloat(process.env.OLLAMA_MODEL_TEMPERATURE);
     }
 
     async prompt(prompt) {
@@ -13,7 +17,7 @@ class OllamaService {
                 model: this.model,
                 prompt: prompt,
                 stream: false,
-                options: { "temperature": 0.0, "top_p": 1.0, "num_ctx": 50, "stop": ["<think></think>"] }
+                options: { "temperature": this.temperature, "top_p": 1.0, "num_ctx": this.maxInputTokenLength, "num_predict": this.maxOutputTokenLength, "stop": ["<think></think>"] }
             });
 
             const message = replaceThinkTags(response.response);
@@ -24,13 +28,16 @@ class OllamaService {
         }
     }
 
-    async chat(messages) {
+    async chat(messageHistory) {
+        let messages = createRoleMessages(messageHistory);
+        messages = createSystemMessage(messages, this.defaultPrompt);
+
         try {
             const response = await this.networkManager.request('/api/chat', 'POST', {
                 model: this.model,
                 messages: messages,
                 stream: false,
-                options: { "temperature": 0.0, "top_p": 1.0, "num_ctx": 50, "stop": ["<think></think>"] }
+                options: { "temperature": this.temperature, "top_p": 1.0, "num_ctx": this.maxInputTokenLength, "num_predict": this.maxOutputTokenLength, "stop": ["<think></think>"] }
             });
 
             const message = replaceThinkTags(response.message.content);
@@ -47,4 +54,20 @@ module.exports = ollamaService;
 
 function replaceThinkTags(message) {
     return message.replace(/[`\s]*[\[\<]think[\>\]](.*?)[\[\<]\/think[\>\]][`\s]*|^[`\s]*([\[\<]thinking[\>\]][`\s]*.*)$/ims, '');
+}
+
+function createRoleMessages(messages) {
+    return messages.map(m => {
+        if (m.author.bot) {
+            return { 'role': 'assistant', 'content': `${m.embeds[0].description}` };
+        } else {
+            return { 'role': 'user', 'content': `${m.content}` };
+        }
+
+    }).reverse();
+}
+
+function createSystemMessage(messages, message) {
+    messages.unshift({ 'role': 'system', 'content': message });
+    return messages;
 }
